backend/requirements.txt
fastapi
uvicorn
moviepy
pillow
python-dotenv